requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
django-celery-beat==2.7.0
pillow==11.0.0
psutil==6.1.1
python-dotenv==1.0.1
sqlparse==0.5.3
channels[daphne]==4.2.0
celery==5.4.0
redis==5.2.1
django-redis==5.4.0
channels_redis==4.2.1
smartpasslib==1.2.0
django-tinymce==4.1.0
psycopg2-binary==2.9.10
django-debug-toolbar==6.0.0